LPS-80123 (portal-tools-service-builder) Do not rewrite extending UADExporter class

+++ modules/util/portal-tools-service-builder/src/main/java/com/liferay/portal/tools/service/builder/ServiceBuilder.java
@@ -4175,8 +4175,10 @@ public class ServiceBuilder {
 				_uadOutputPath, "/uad/exporter/", entity.getName(),
 				"UADExporter.java"));
 
-		ToolsUtil.writeFile(
-			file, content, _author, _jalopySettings, _modifiedFileNames);
+		if (!file.exists()) {
+			ToolsUtil.writeFile(
+				file, content, _author, _jalopySettings, _modifiedFileNames);
+		}
 	}
 
 	private void _createUADExporterTest(Entity entity) throws Exception {
